requirements.txt
Pillow
ruff
numpy
opencv-python
